+++ mypyc/emitmodule.py
@@ -36,7 +36,7 @@ from mypyc.ir.ops import LiteralsMap, DeserMaps
 from mypyc.rtypes import RType, RTuple
 from mypyc.ir.func_ir import FuncIR
 from mypyc.ir.class_ir import ClassIR
-from mypyc.module_ir import ModuleIR, ModuleIRs, deserialize_modules
+from mypyc.ir.module_ir import ModuleIR, ModuleIRs, deserialize_modules
 from mypyc.options import CompilerOptions
 from mypyc.uninit import insert_uninit_checks
 from mypyc.refcount import insert_ref_count_opcodes
