+++ mypyc/test/test_serialization.py
@@ -11,7 +11,7 @@ from mypyc.ir.ops import DeserMaps
 from mypyc.rtypes import RType
 from mypyc.ir.func_ir import FuncDecl, FuncIR, FuncSignature
 from mypyc.ir.class_ir import ClassIR
-from mypyc.module_ir import ModuleIR, deserialize_modules
+from mypyc.ir.module_ir import ModuleIR, deserialize_modules
 from mypyc.sametype import is_same_type, is_same_signature
